emd.cpp updated
Add comments to increase understanding 

Improve function and variable naming convention in recent standards

Add error checking: The code assumes that the input tensors are of the correct shape and type. However, it does not check this assumption. Adding error checking will help catch mistakes early.

Use const where appropriate: The input tensors are not modified by the functions. Therefore, they should be declared as const to indicate that they are read-only.

Use references instead of pointers: The input tensors are passed as pointers. However, it would be better to pass them as references to make the code more readable and avoid the need for pointer dereferencing.